Remove grok from opengever.repository.
