+++ opengever/repository/browser/configure.zcml
@@ -53,4 +53,35 @@
       permission="cmf.ModifyPortalContent"
       />
 
+  <browser:page
+      for="opengever.repository.repositoryfolder.IRepositoryFolderSchema"
+      name="delete_repository"
+      class=".deletion.RepositoryDeletionView"
+      permission="zope2.DeleteObjects"
+      template="templates/deletion.pt"
+      />
+
+  <browser:page
+      for="*"
+      name="is_deletion_allowed"
+      class=".deletion.RepositoryDeletionAllowed"
+      permission="zope2.View"
+      />
+
+  <browser:page
+      for="opengever.repository.repositoryroot.IRepositoryRoot"
+      name="tabbedview_view-overview"
+      class=".overview.RepositoryRootOverview"
+      template="templates/repositoryrootoverview.pt"
+      permission="zope2.View"
+      />
+
+  <browser:page
+      for="*"
+      name="primary_repository_root"
+      class=".primary_repository_root.PrimaryRepositoryRoot"
+      template="templates/repositoryrootoverview.pt"
+      permission="zope2.View"
+      />
+
 </configure>
