+++ opengever/repository/browser/templates/deletion.pt
