+++ opengever/repository/indexers.py
@@ -1,4 +1,3 @@
-from five import grok
 from opengever.repository.repositoryfolder import IRepositoryFolder
 from plone.indexer import indexer
 
@@ -7,11 +6,7 @@ from plone.indexer import indexer
 def title_de_indexer(obj):
     return obj.get_prefixed_title_de()
 
-grok.global_adapter(title_de_indexer, name="title_de")
-
 
 @indexer(IRepositoryFolder)
 def title_fr_indexer(obj):
     return obj.get_prefixed_title_fr()
-
-grok.global_adapter(title_fr_indexer, name="title_fr")
